# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /crm/funnels:
    get:
      operationId: listCrmFunnels
      tags: [crm]
      summary: List CRM funnels
      responses:
        "200":
          description: CRM funnels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrmFunnel"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCrmFunnel
      tags: [crm]
      summary: Create a CRM funnel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmFunnelInput"
      responses:
        "201":
          description: Created CRM funnel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmFunnel"
        "400":
          description: Invalid funnel data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCrmFunnel
      tags: [crm]
      summary: Update a CRM funnel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmFunnelUpdate"
      responses:
        "200":
          description: Updated CRM funnel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmFunnel"
        "400":
          description: Invalid funnel update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Funnel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /crm/funnels/{id}/cost-history:
    get:
      operationId: listCrmFunnelCostHistory
      tags: [crm]
      summary: List a CRM funnel cost history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Funnel cost history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrmFunnelCostHistory"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /crm/ads:
    get:
      operationId: listCrmAds
      tags: [crm]
      summary: List CRM ads
      responses:
        "200":
          description: CRM ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrmAd"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /crm/ads/unlinked:
    get:
      operationId: listUnlinkedCrmAds
      tags: [crm]
      summary: List CRM ads without a funnel
      responses:
        "200":
          description: Unlinked CRM ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrmAd"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /crm/ads/{id}:
    patch:
      operationId: linkCrmAd
      tags: [crm]
      summary: Link or unlink a CRM ad from a funnel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmAdLink"
      responses:
        "200":
          description: Updated CRM ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmAd"
        "400":
          description: Invalid ad link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Ad not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CrmFunnel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        current_cost_per_lead:
          type: [string, "null"]
        cost_updated_at:
          type: [string, "null"]
          format: date-time
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, is_active, created_at, updated_at]
    CrmFunnelInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        current_cost_per_lead:
          type: [string, "null"]
        is_active:
          type: boolean
      required: [name, current_cost_per_lead]
    CrmFunnelUpdate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          minLength: 1
        current_cost_per_lead:
          type: [string, "null"]
        is_active:
          type: boolean
      required: [id]
    CrmFunnelCostHistory:
      type: object
      properties:
        id:
          type: string
        funnel_id:
          type: string
        cost_per_lead:
          type: string
        valid_from:
          type: string
          format: date
        valid_to:
          type: [string, "null"]
          format: date
        updated_by:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
      required: [id, funnel_id, cost_per_lead, valid_from, created_at]
    CrmAd:
      type: object
      properties:
        id:
          type: string
        facebook_ad_id:
          type: string
        name:
          type: [string, "null"]
        ad_url:
          type: [string, "null"]
        funnel_id:
          type: [string, "null"]
        last_synced_at:
          type: [string, "null"]
          format: date-time
        fetch_failed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, facebook_ad_id, fetch_failed, created_at, updated_at]
    CrmAdLink:
      type: object
      properties:
        funnel_id:
          type: [string, "null"]
      required: [funnel_id]
...