# lib/api-zod/src/generated/types/crmFunnelCostHistory.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CrmFunnelCostHistory {
  id: string;
  funnel_id: string;
  cost_per_lead: string;
  valid_from: Date;
  /** @nullable */
  valid_to?: Date | null;
  /** @nullable */
  updated_by?: string | null;
  created_at: Date;
}