# lib/api-zod/src/generated/types/crmFunnelUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CrmFunnelUpdate {
  id: string;
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  current_cost_per_lead?: string | null;
  is_active?: boolean;
}